requirements.txt
fastapi==0.104.1
openai==1.3.5
qdrant-client==1.8.0
psycopg2-binary==2.9.9
pydantic==1.10.13
python-dotenv==1.0.0
uvicorn==0.24.0
pytest==7.4.3